Make CXFA_List::GetLength return a size_t
This requires a number of other minor cleanups to get sizes and
signedness agreeing everywhere.

+++ xfa/fxfa/parser/cxfa_node.h
@@ -153,7 +153,7 @@ class CXFA_Node : public CXFA_Object {
   CXFA_Node* GetModelNode();
   void UpdateNameHash();
 
-  int32_t CountChildren(XFA_Element eType, bool bOnlyChild);
+  size_t CountChildren(XFA_Element eType, bool bOnlyChild);
 
   template <typename T>
   T* GetChild(int32_t index, XFA_Element eType, bool bOnlyChild) {
